refactor(runtime): separate context preflight and activation effects
